# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Website Builder Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface RegisterBody {
  fullName: string;
  phone: string;
  email: string;
  username: string;
  password: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface User {
  id: number;
  fullName: string;
  phone: string;
  email: string;
  username: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

/**
 * @nullable
 */
export type OrderPackage = {
  id?: number;
  name?: string;
  priceEgp?: number;
  priceSar?: number;
} | null;

/**
 * @nullable
 */
export type OrderPaymentMethod = {
  id?: number;
  name?: string;
  details?: string;
} | null;

export interface Order {
  id: number;
  userId: number;
  siteName: string;
  siteType: string;
  details: string;
  /** @nullable */
  packageId?: number | null;
  /** @nullable */
  customBudget?: number | null;
  currency: string;
  /** @nullable */
  paymentMethodId?: number | null;
  status: string;
  depositPaid: boolean;
  finalPaid: boolean;
  /** @nullable */
  totalAmount?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
  user: User;
  /** @nullable */
  package?: OrderPackage;
  /** @nullable */
  paymentMethod?: OrderPaymentMethod;
}

export interface CreateOrderBody {
  siteName: string;
  siteType: string;
  details: string;
  /** @nullable */
  packageId?: number | null;
  /** @nullable */
  customBudget?: number | null;
  currency: string;
  /** @nullable */
  paymentMethodId?: number | null;
}

export interface UpdateOrderBody {
  /** @nullable */
  status?: string | null;
  /** @nullable */
  depositPaid?: boolean | null;
  /** @nullable */
  finalPaid?: boolean | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  totalAmount?: number | null;
}

export interface Package {
  id: number;
  name: string;
  description: string;
  priceEgp: number;
  priceSar: number;
  features: string;
  isActive: boolean;
  createdAt: string;
}

export interface CreatePackageBody {
  name: string;
  description: string;
  priceEgp: number;
  priceSar: number;
  features: string;
  isActive?: boolean;
}

export interface UpdatePackageBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  priceEgp?: number | null;
  /** @nullable */
  priceSar?: number | null;
  /** @nullable */
  features?: string | null;
  /** @nullable */
  isActive?: boolean | null;
}

export interface PaymentMethod {
  id: number;
  name: string;
  details: string;
  isActive: boolean;
  createdAt: string;
}

export interface CreatePaymentMethodBody {
  name: string;
  details: string;
  isActive?: boolean;
}

export interface UpdatePaymentMethodBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  isActive?: boolean | null;
}

export interface Testimonial {
  id: number;
  clientName: string;
  comment: string;
  rating: number;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateTestimonialBody {
  clientName: string;
  comment: string;
  rating: number;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
}

export interface UpdateTestimonialBody {
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  comment?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  isActive?: boolean | null;
}

export interface AdminStats {
  totalUsers: number;
  totalOrders: number;
  pendingOrders: number;
  completedOrders: number;
  inProgressOrders: number;
  totalRevenue: number;
  recentOrders: Order[];
}

export type ListOrdersParams = {
  /**
   * @nullable
   */
  userId?: number | null;
  /**
   * @nullable
   */
  status?: string | null;
};