# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Website Builder Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin) or user's orders
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new website order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List all packages
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"

    post:
      operationId: createPackage
      tags: [packages]
      summary: Create package (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageBody"
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{id}:
    patch:
      operationId: updatePackage
      tags: [packages]
      summary: Update package (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePackageBody"
      responses:
        "200":
          description: Updated package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"

    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete package (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /payment-methods:
    get:
      operationId: listPaymentMethods
      tags: [payment-methods]
      summary: List all payment methods
      responses:
        "200":
          description: List of payment methods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethod"

    post:
      operationId: createPaymentMethod
      tags: [payment-methods]
      summary: Create payment method (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentMethodBody"
      responses:
        "201":
          description: Payment method created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
  /payment-methods/{id}:
    patch:
      operationId: updatePaymentMethod
      tags: [payment-methods]
      summary: Update payment method (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentMethodBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"

    delete:
      operationId: deletePaymentMethod
      tags: [payment-methods]
      summary: Delete payment method (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        username:
          type: string
        password:
          type: string
      required:
        - fullName
        - phone
        - email
        - username
        - password
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        username:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required:
        - id
        - fullName
        - phone
        - email
        - username
        - role
        - createdAt
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        siteName:
          type: string
        siteType:
          type: string
        details:
          type: string
        packageId:
          type: ["integer", "null"]
        customBudget:
          type: ["number", "null"]
        currency:
          type: string
        paymentMethodId:
          type: ["integer", "null"]
        status:
          type: string
        depositPaid:
          type: boolean
        finalPaid:
          type: boolean
        totalAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
        package:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            name:
              type: string
            priceEgp:
              type: number
            priceSar:
              type: number
        paymentMethod:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            name:
              type: string
            details:
              type: string
      required:
        - id
        - userId
        - siteName
        - siteType
        - details
        - currency
        - status
        - depositPaid
        - finalPaid
        - createdAt
        - updatedAt
        - user
    CreateOrderBody:
      type: object
      properties:
        siteName:
          type: string
        siteType:
          type: string
        details:
          type: string
        packageId:
          type: ["integer", "null"]
        customBudget:
          type: ["number", "null"]
        currency:
          type: string
        paymentMethodId:
          type: ["integer", "null"]
      required:
        - siteName
        - siteType
        - details
        - currency
    UpdateOrderBody:
      type: object
      properties:
        status:
          type: ["string", "null"]
        depositPaid:
          type: ["boolean", "null"]
        finalPaid:
          type: ["boolean", "null"]
        notes:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
    Package:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        priceEgp:
          type: number
        priceSar:
          type: number
        features:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - priceEgp
        - priceSar
        - features
        - isActive
        - createdAt
    CreatePackageBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        priceEgp:
          type: number
        priceSar:
          type: number
        features:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - description
        - priceEgp
        - priceSar
        - features
    UpdatePackageBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        priceEgp:
          type: ["number", "null"]
        priceSar:
          type: ["number", "null"]
        features:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    PaymentMethod:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        details:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - details
        - isActive
        - createdAt
    CreatePaymentMethodBody:
      type: object
      properties:
        name:
          type: string
        details:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - details
    UpdatePaymentMethodBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        details:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
...