# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Website Builder Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderBody {
  siteName: string;
  siteType: string;
  details: string;
  /** @nullable */
  packageId?: number | null;
  /** @nullable */
  customBudget?: number | null;
  currency: string;
  /** @nullable */
  paymentMethodId?: number | null;
}