# lib/api-zod/src/generated/types/createPackageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Website Builder Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePackageBody {
  name: string;
  description: string;
  priceEgp: number;
  priceSar: number;
  features: string;
  isActive?: boolean;
}