# lib/api-zod/src/generated/types/createPaymentMethodBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Website Builder Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePaymentMethodBody {
  name: string;
  details: string;
  isActive?: boolean;
}