# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Website Builder Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPackage } from "./orderPackage";
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { User } from "./user";

export interface Order {
  id: number;
  userId: number;
  siteName: string;
  siteType: string;
  details: string;
  /** @nullable */
  packageId?: number | null;
  /** @nullable */
  customBudget?: number | null;
  currency: string;
  /** @nullable */
  paymentMethodId?: number | null;
  status: string;
  depositPaid: boolean;
  finalPaid: boolean;
  /** @nullable */
  totalAmount?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
  user: User;
  /** @nullable */
  package?: OrderPackage;
  /** @nullable */
  paymentMethod?: OrderPaymentMethod;
}