# lib/api-zod/src/generated/types/paymentMethod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Website Builder Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentMethod {
  id: number;
  name: string;
  details: string;
  isActive: boolean;
  createdAt: string;
}