# lib/api-zod/src/generated/types/updatePackageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Website Builder Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePackageBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  priceEgp?: number | null;
  /** @nullable */
  priceSar?: number | null;
  /** @nullable */
  features?: string | null;
  /** @nullable */
  isActive?: boolean | null;
}