# lib/api-zod/src/generated/types/updatePaymentMethodBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Website Builder Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePaymentMethodBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  isActive?: boolean | null;
}