# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Website Builder Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  fullName: string;
  phone: string;
  email: string;
  username: string;
  role: string;
  createdAt: string;
}